move iterator tools to dedicated module
